refactor(analysis): Phase 6 S5a — relocate TransitionalScheduleRunner -> analysis/transition.py
Pure move (behaviour-preserving): TransitionalScheduleRunner + _extract_floor_metrics
+ _build_timeline_lazyframe + _TRANSITIONAL_REPORTING_DATES moved verbatim from
analysis/comparison.py into a new analysis/transition.py. Repointed the two test
files (test_transitional_schedule, test_scenario_m33) and the comparison.py module
docstring; dropped the now-unused PermissionMode / TransitionalScheduleBundle
imports from comparison.py. Still returns TransitionalScheduleBundle (contracts/,
unchanged), still N full pipeline runs per year.

S5b replaces the per-year full-pipeline loop with a floor-tail partial re-run,
behind a partial-vs-full equivalence test.

Gate: ruff + ty clean; arch_check EXIT 0; full suite 7514 passed / 2 skipped.
